+++ measures/nrc_report_setpointdiff/resources/NRCReportingMeasureHelper.rb
@@ -13,15 +13,29 @@ module NRCReportingMeasureTestHelper
 
   # Define the output path. Set defaults and remove any existing outputs.
   @output_root_path = File.expand_path("#{File.expand_path(__dir__)}/../tests/output")
-  Dir.mkdir @output_root_path unless Dir.exists?(@output_root_path)
+  Dir.mkdir @output_root_path unless Dir.exists?(@output_root_path) 
   @output_path = @output_root_path
-  existing_folders = Dir.entries(@output_root_path) - ['.', '..'] # Remove current folder above from list before deleting!
-  existing_folders.each do |entry|
-    folder_to_remove = File.expand_path("#{@output_root_path}/#{entry}")
-    puts "Removing existing output folder: #{folder_to_remove}".yellow
-    FileUtils.rm_rf(folder_to_remove)
-  end
 
+  # Remove the existing test results. Need to control when this is done as multiple test scripts could be
+  #  accessing the same path.
+  # Must call this in the test script.
+  def self.removeOldOutputs(before: Time.now)
+    existing_folders = Dir.entries(@output_path) - ['.', '..'] # Remove current folder above from list before deleting!
+    existing_folders.each do |entry|
+      folder_to_remove = File.expand_path("#{@output_path}/#{entry}")
+	  if (Dir.exist?(folder_to_remove)) # Double check it exists (incase another process has removed it as is the case with multiple test files).
+        puts "Checking existing output folder: #{before}; #{File.mtime(folder_to_remove)}; #{folder_to_remove}".green
+        if File.mtime(folder_to_remove) < before
+          puts "Removing folder: #{folder_to_remove}".yellow
+          FileUtils.rm_rf(folder_to_remove)
+        else
+          puts "Skipping existing output folder: #{folder_to_remove}".light_blue
+        end
+      end
+    end
+  end
+  
+  #
   # Define methods to manage output folders.
   def self.resetOutputFolder
     @output_path = @output_root_path
@@ -29,8 +43,8 @@ module NRCReportingMeasureTestHelper
 
   def self.appendOutputFolder(folder)
     # Append name and validate if specified by the user
-    path = @output_root_path + "/" + folder
-    validateOutputFolder(path)
+    path = @output_path + "/" + folder
+  validateOutputFolder(path)
   end
 
   def self.validateOutputFolder(path)
@@ -38,9 +52,10 @@ module NRCReportingMeasureTestHelper
     # Also check if it exists.
     # By default use the test method name.
     path = File.expand_path(path)
-    if path == @output_root_path
+    if path == @output_root_path 
       # Append the calling method name and re-validate (need to jump back two methods)
-      path = @output_root_path + "/" + caller_locations(1, 2)[1].label.split.last
+      path = @output_root_path + "/" + caller_locations(1,2)[1].label.split.last
+	  puts "Appending path to test output folder: #{path}"
       validateOutputFolder(path)
     elsif File.exist?(path)
       # Create a numbered subfolder. First check if there is a numbered folder.
@@ -58,34 +73,65 @@ module NRCReportingMeasureTestHelper
     @output_path.to_s
   end
 
+  #
+  # Define the folder containing the test file for the test case summary output file.
+  @measure_path = File.expand_path("#{File.expand_path(__dir__)}/../tests")
+  @test_summary_mdfile = @measure_path + "/" + "README.md"
+  FileUtils.rm_rf(@test_summary_mdfile)
+  
+  # Create initial file with title.
+  @testSummaryTitleRequired = true
+  def self.testSummaryTitleRequired
+    return @testSummaryTitleRequired
+  end
+  
+  def self.setTestSummaryTitle(required)
+    @testSummaryTitleRequired = required
+  end
+  
+  def self.testSummaryMDfile
+    @test_summary_mdfile.to_s
+  end
+  
+  # Test count
+  @testSummaryCount = 1
+  def self.incrementTestSummaryCount
+    @testSummaryCount += 1
+  end
+  
+  def self.testSummaryCount
+    return @testSummaryCount
+  end
+
+  #
   # Custom way to run a reporting measure in the test. Overwrites run_measure definition in BTAPMeasureTestHelper.
   def run_measure(input_arguments, model)
-
+  
     # Provide feedback as to what is being done to teh terminal.
     puts "Running measure".green
     puts "  with input arguments".green + " #{input_arguments}".light_blue
     puts "  on model with".green + " #{model.modelObjects.count}".light_blue + " objects".green
-    puts "  from method".green + " #{caller_locations(1, 1)[0].label.split.last}".light_blue
+    puts "  from method".green + " #{caller_locations(1,1)[0].label.split.last}".light_blue
 
     # Set the output folder. This should be unique (check done in validateOutputFolder). Create if does not exist.
     output_folder = NRCReportingMeasureTestHelper.outputFolder
     output_folder = NRCReportingMeasureTestHelper.validateOutputFolder(output_folder)
-    Dir.mkdir(output_folder) unless Dir.exists?(output_folder)
-
+    FileUtils.mkdir_p(output_folder) unless Dir.exists?(output_folder)
+  
     # This will create a instance of the measure you wish to test. It does this based on the test class name.
     measure = get_measure_object()
     measure.use_json_package = @use_json_package
     measure.use_string_double = @use_string_double
-
+  
     # Return false if can't
     return false if false == measure
-
+  
     # Now get the arguments and create a runner
     arguments = measure.arguments()
     argument_map = OpenStudio::Measure.convertOSArgumentVectorToMap(arguments)
     runner = OpenStudio::Measure::OSRunner.new(OpenStudio::WorkflowJSON.new)
-
-    # Set the arguments in the argument map use json or real arguments.
+  
+    # Set the arguements in the argument map use json or real arguments.
     if @use_json_package
       argument = arguments[0].clone
       assert(argument.setValue(input_arguments['json_input']), "Could not set value for 'json_input' to #{input_arguments['json_input']}")
@@ -102,6 +148,9 @@ module NRCReportingMeasureTestHelper
         argument_map[key] = argument
       end
     end
+  
+    # Get the e+ output requests, this will be done automatically by OS App and PAT
+    idf_output_requests = measure.energyPlusOutputRequests(runner, argument_map)
 
     # Mimic the process of running this measure in OS App or PAT.
     # Create a new folder to work in (if it does not exist)
@@ -122,17 +171,7 @@ module NRCReportingMeasureTestHelper
       FileUtils.rm(model_out_path)
     end
 
-    # Set up runner
-    runner.setLastOpenStudioModel(model)
-
-    # get arguments
-    arguments = measure.arguments()
-    argument_map = OpenStudio::Measure.convertOSArgumentVectorToMap(arguments)
-
-    # Get the e+ output requests, this will be done automatically by OS App and PAT
-    idf_output_requests = measure.energyPlusOutputRequests(runner, argument_map)
-
-    # Convert output requests to OSM for testing, OS App and PAT will add these to the E+ Idf
+    # Convert output requests to OSM for testing, OS App and PAT will add these to the E+ Idf.
     workspace = OpenStudio::Workspace.new('Draft'.to_StrictnessLevel, 'EnergyPlus'.to_IddFileType)
     workspace.addObjects(idf_output_requests)
     rt = OpenStudio::EnergyPlus::ReverseTranslator.new
@@ -141,15 +180,14 @@ module NRCReportingMeasureTestHelper
     # Add requested outputs to model.
     model.addObjects(request_model.objects)
     model.save(model_out_path, true)
-
+    
     sql_path = "#{output_folder}/run/eplusout.sql"
-
     if ENV['OPENSTUDIO_TEST_NO_CACHE_SQLFILE']
       if File.exist?(sql_path)
         FileUtils.rm_f(sql_path)
       end
     end
-
+    
     osw_path = File.join(output_folder, 'in.osw')
     osw_path = File.absolute_path(osw_path.to_s)
 
@@ -163,14 +201,12 @@ module NRCReportingMeasureTestHelper
     cmd = "\"#{cli_path}\" run -w \"#{osw_path}\""
     puts "Running openstudio with command: ".green + "\n#{cmd}".light_blue
     OpenstudioStandards.run_command(cmd)
-
-    assert(File.exist?(model_out_path), "Could not find osm at this path:#{model_out_path}".red)
-    assert(File.exist?(sql_path), "Could not find sql at this path:#{sql_path}".red)
-
-    # set up runner, this will happen automatically when measure is run in PAT or OpenStudio
+  
+    # Set up runner, this will happen automatically when measure is run in PAT or OpenStudio.
     runner.setLastOpenStudioModelPath(model_out_path)
     runner.setLastEpwFilePath(epw_path)
     runner.setLastEnergyPlusSqlFilePath(sql_path)
+  
     # Temporarily change directory to the run directory and run the measure
     start_dir = Dir.pwd
     begin
@@ -179,14 +215,51 @@ module NRCReportingMeasureTestHelper
     ensure
       Dir.chdir(start_dir)
     end
-
-    runner.result
+  
     # Reset the output path to the root folder.
     NRCReportingMeasureTestHelper.resetOutputFolder
+
+    # Add summary of test to README file
+    measure_name = measure.name.gsub("_", " ").upcase
+    reportCase(measure_name, output_folder.split('/').last, input_arguments)
+
     return runner
   end
 
-  #Fancy way of getting the measure object automatically. Added check for NRC in measure name.
+  # Method to report case being tested
+  def reportCase(measure_name, test_name, input_arguments)
+
+    # File name defined above. Open for appending.
+    out_file = File.new("#{NRCReportingMeasureTestHelper.testSummaryMDfile}", "a")
+
+    # Only add the page title once.
+    if NRCReportingMeasureTestHelper.testSummaryTitleRequired
+      title = "# Summary Of Test Cases for '#{measure_name}' Measure"
+      out_file.puts("#{title}")
+      out_file.puts(" ")
+    end
+  
+    # Current test name.
+    test_name = test_name.gsub("_", " ")
+    out_file.puts("## #{NRCReportingMeasureTestHelper.testSummaryCount} - #{test_name}")
+  
+    # Create a table describing the case tested. Table header first.
+    out_file.puts("| Test Argument | Test Value |")
+    out_file.puts("| ------------- | ---------- |")
+
+    # Table contents.
+    input_arguments.each do |key, value|
+      out_file.puts("| #{key} |#{value} |")
+    end
+    out_file.puts(" ")
+    out_file.close
+  
+    # Update logical and counters.
+    NRCReportingMeasureTestHelper.setTestSummaryTitle(false)
+    NRCReportingMeasureTestHelper.incrementTestSummaryCount
+  end
+
+  # Fancy way of getting the measure object automatically. Added check for NRC in measure name.
   def get_measure_object()
     measure_class_name = self.class.name.to_s.match((/(NRC.*)(\_Test)/i) || ((/(BTAP.*)(\_Test)/i))).captures[0]
     btap_measure = nil
@@ -207,12 +280,31 @@ module NRCReportingMeasureTestHelper
       return nrc_measure
     end
   end
+
+  # Load a test model (code that is common in a lot of test scripts). Returns the model object.
+  def load_test_osm(full_osm_model_path)
+  
+    # Load the supplied osm.
+    translator = OpenStudio::OSVersion::VersionTranslator.new
+    path = OpenStudio::Path.new(full_osm_model_path)
+    model = translator.loadModel(path)
+    assert((not model.empty?), "Reading model file: #{path}")
+    model = model.get
+  end
 end
 
-# Add significant digits capability to float class.
+# Add significant digits capability to float amd integer class.
 class Float
-  def signif(digits = 3)
+  def signif(digits=3)
     return 0 if self.zero?
+    return self if self < 0.0
     self.round(-(Math.log10(self).ceil - digits))
+  end
+end
+class Integer
+  def signif(digits=3)
+    return 0 if self.zero?
+    return self if self < 0
+    self.round(-(Math.log10(self).ceil - digits)).to_i
   end
 end
